+++ tests/test_simulations.py
@@ -233,20 +233,14 @@ class TestSimulation():
         # Check bad grid
         hx = np.ones(17)*20
         grid = emg3d.TensorMesh([hx, hx, hx], (0, 0, 0))
-        print(80*'*')
-        print(hx)
         with pytest.warns(UserWarning, match='optimal for MG solver. Good n'):
             simulations.Simulation(self.survey, self.model, gridding='input',
                                    gridding_opts=grid)
 
     def test_synthetic(self):
         sim = self.simulation.copy()
-
-        # Switch off noise_floor, relative_error, min_offset => No noise.
-        sim.survey.noise_floor = None
-        sim.survey.relative_error = None
         sim._dict_efield = sim._dict_initiate  # Reset
-        sim.compute(observed=True)
+        sim.compute(observed=True, add_noise=False)
         assert_allclose(sim.data.synthetic, sim.data.observed)
         assert sim.survey.size == sim.data.observed.size
 
@@ -262,7 +256,8 @@ class TestSimulation():
                              'linerelaxation': False, 'semicoarsening': False},
                 gridding='input', gridding_opts=newgrid, name='TestX')
 
-        grad = simulation.gradient
+        with pytest.warns(UserWarning, match='Receiver responses were obtain'):
+            grad = simulation.gradient
 
         # Ensure the gradient has the shape of the model, not of the input.
         assert grad.shape == self.model.shape
@@ -426,6 +421,22 @@ class TestSimulation():
         simulation.compute()
         out, _ = capsys.readouterr()
         assert "= Source TxED-1; Frequency 1.0 Hz = CONVERGED" in out
+
+        # Two sources, only compute 1, assure printing works.
+        sources = [emg3d.TxElectricDipole((x, 0, 0, 0, 0)) for x in [0, 10]]
+        survey = emg3d.Survey(
+            name='Test', sources=sources,
+            receivers=receivers,
+            frequencies=1.0, noise_floor=1e-15, relative_error=0.05,
+        )
+
+        inp = {'name': 'Test', 'survey': survey, 'model': model,
+               'gridding': 'same'}
+        simulation = simulations.Simulation(**inp, solver_opts={'verb': 0})
+        _ = simulation.get_efield('TxED-2', 'f-1')
+        simulation.print_solver_info(verb=1)
+        out, _ = capsys.readouterr()
+        assert "= Source TxED-2; Frequency 1.0 Hz = CONVERGED" in out
 
     def test_rel_abs_rec(self):
         # Sources